added createContact to contactList
